Use MemoryStore package
The MemoryStore implementation included in express-sssion is not
suitable for production and outputs a warning. Instead we the
MemoryStore package. However, at least when running unit tests
this doesn't properly shut down, therefore we use it only in
production. We'll have to see if this is suitable on the production
server.