Explicitly add webdriver-manager
protractor depends on webdriver-manager, but the version it depends
on has a bug that downloads an incompatible version of Chrome
which causes the build/tests to fail. As a workaround we explicitly
add a version of webdriver-manager that fixes this problem
(see https://stackoverflow.com/questions/55910242/when-using-\
protractor-5-4-2-webdriver-manager-downloads-2-46-which-is-not-compa).