Use MemoryStore package (#25)
The MemoryStore implementation included in express-session is not
suitable for production and outputs a warning. Instead we use the
MemoryStore package. However, at least when running unit tests
this doesn't properly shut down, therefore we don't use it when
running tests. We'll have to see if this is suitable on the
production server.